Change CI configuration (#145)
The changes in this PR try to enforce the use of bors to a larger extent. PRs are not tested automatically anymore but instead only when one comments with `bors try` (does not merge afterwards) or `bors r+` (merges automatically afterwards). Additionally, the configuration requires at least one approval of a project member and squashes commits before merging.

I'm not sure about building the master branch as well - in my experience it leads to redundant tests (since the commits are tested in the staging branch already) but works a bit better with the display of the Github check status and the badges.