Added debug statements to link! and invlink! (#206)
As mentioned here: https://github.com/TuringLang/Turing.jl/pull/1497/files#r547599474

Figured it would be best to make a PR and discuss whether this change is desirable here. Are there any perf-considerations to this? We already have if-statements in there, e.g. `istrans`, so another one due to the `@debug` shouldn't make much of a difference.