Fixes for Julia 1.6 (#212)
This PR fixes https://github.com/TuringLang/Turing.jl/issues/1525#issuecomment-760900398 and fixes the model definition and some warning on Julia 1.6. It also includes a line number node of the call site instead of a reference to `src/compiler.jl` in the model definition for easier debugging. On Julia 1.6, initially I got warnings about method redefinitions (models of the same name) that all referred to line 344 in `src/compiler.jl` instead of the place where the model was defined by the user.

Turing tests are disabled currently since they crash (segmentation fault). Might be related to the fact that Libtask is not compatible with Julia 1.6 yet (therefore also an older version of Turing and Libtask would be used on Julia 1.6).

Co-authored-by: David Widmann <devmotion@users.noreply.github.com>