Small addition to `link!` and `invlink!` (#231)
Makes it possible to only link/invlink a subset of the space of a sampler *for static models only* (as there's no equivalent for `UntypedVarInfo`). Is non-breaking since this is just adding an intermediate method to an existing implementation, giving increased flexbility.

Probably don't want to encourage use of this, but it can be useful in cases such as the MH sampler https://github.com/TuringLang/Turing.jl/pull/1582#issuecomment-817148192.