Fix use of arrays of Distributions (#245)
This PR fixes https://github.com/TuringLang/DynamicPPL.jl/issues/28#issuecomment-829223577 and allows to use arbitrary arrays of `Distribution`s. This was already allowed in the context implementations but prevented by a check in the code generated by the `@model` macro.

Additionally, the PR replaces the hard-coded check with a `check_tilde_rhs` function which, IMO, makes the code a bit simpler and easier to read. Moreover, a bug in the `dot_assume` implementation for arrays of Distributions is fixed.

Co-authored-by: David Widmann <devmotion@users.noreply.github.com>