Removal of redundant backwards compat `to_namedtuple_expr` (#280)
Looking at `to_namedtuple_expr` it seems like it's leftover code to be compatible with Julia 1.2. Given that DPPL only supports Julia 1.3 or higher I've simplified the implementation, doing away with the `namedtuple` method.

EDIT: This also fixed a bug I ran into when using Zygote + `@submodel` in #267 . Struggling to come up with a MWE of the bug, but just adding merging this fixed it. This was also my original motivation for making this change, as Zygote was complaining about the `namedtuple` for some reason (something about "non-differentiable getfield").

Co-authored-by: Hong Ge <hg344@cam.ac.uk>