Reduce CI (#317)
Similar to what we decided to use for JuliaGaussianProcesses. Tests with Julia 1.3 are important since we state that DynamicPPL is compatible with 1.3 and it is the only way that we discover if we break something by using e.g. language features that are not available on Julia 1.3 before making a release. However, I think it should be sufficient to perform these tests with ubuntu. Also I think it is sufficient to test macOS and Windows with Julia 1 with `JULIA_NUM_THREADS = 2`.