Fix for #371 (#372)
This PR adds a method called `resolve_varnames(varname, dist)` and adds an additional generated variable for each `~` which now holds the RHS of `~`. 

It does address #371 but uncertain if this is the best way, so wouldn't recommend merging this just yet. But putting it here so we can colab on it.

Co-authored-by: Hong Ge <3279477+yebai@users.noreply.github.com>