Condition with `Dict` as underlying storage (#419)
This PR allows usage of `Dict` as the underlying storage in addition to the currently supported `NamedTuple`.

Similarly to `SimpleVarInfo`, this gives us two approaches: one with somewhat limited support, as outlined in the docstring, but with (usually) zero runtime overhead (`NamedTuple`), and one with full support but with runtime overead (`Dict`).