Use `OrderedDict` in `SimpleVarInfo` + improvements and fixes for `values_as` (#420)
We are currently using `Dict` together with `SimpleVarInfo` which leads to inconsistent ordering of the variables vs. `OrderdDict` which, if generated from a `Model`, will preserve the execution order of the model.

In addition, I've fixed some impls for `values_as` + added more better support, in addition to proper testing. Given how it's now better tested + is a nice-to-have feature + will likely see extensive use after #417, it also seems reasonable to export `values_as` from DPPL.

EDIT: This should be merged before #417